+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Security in case of vulnerabilities.
 
 ## [Unreleased]
 
+
+## [25.10.13] - 2025-10-13
+
 ### Added
 - `generate_content` Django mangement command
 
@@ -157,7 +160,8 @@ Security in case of vulnerabilities.
 - First version suitable for public consumption.
 
 
-[Unreleased]: https://github.com/nkantar/Starminder/compare/25.10.12...HEAD
+[Unreleased]: https://github.com/nkantar/Starminder/compare/25.10.13...HEAD
+[25.10.13]: https://github.com/nkantar/Starminder/compare/25.10.12...25.10.13
 [25.10.12]: https://github.com/nkantar/Starminder/compare/25.10.11...25.10.12
 [25.10.11]: https://github.com/nkantar/Starminder/compare/3.0.0...25.10.11
 [3.0.0]: https://github.com/nkantar/Starminder/compare/2.0.0...3.0.0